feat(starry-kmod): support LoongArch DMW-backed kmods
LoongArch kernel symbols are exported from the DMW window, while the
page-table-backed kernel space uses a different high-half address range.
Allocating module sections from the ordinary kernel address space can make
PCALA relocations reconstruct the wrong alias for DMW kernel symbols.

Split kmod section memory into backend-specific handling and allocate
LoongArch module sections from direct-map pages. DMW-backed sections skip
PTE permission updates because DMW translations do not consult page tables,
and release pages through the global allocator instead of kernel_aspace.

Also move the kprobe smoke test into a loadable kmod, expose the kernel kprobe
registration helpers for module use, update kmod-loader and workspace
dependencies, and document the new kprobe_test module support matrix.
