+++ Cargo.toml
@@ -326,3 +326,4 @@ bcm2835-sdhci = "0.1.1-preview.1"
 ixgbe-driver = "0.1.1-preview.1"
 x86 = "0.52"
 kmod-tools = "0.3"
+kprobe = "0.6"
